+++ lib/src/sink/dev_log_sink.dart
@@ -1,20 +1,16 @@
 import 'dart:async';
 import 'dart:developer' as developer;
 
-import 'package:fox_logging/src/filter/log_filter.dart';
 import 'package:fox_logging/src/sink/log_sink.dart';
 import 'package:logging/logging.dart';
 
 /// A [LogSinkMixin] which uses the [developer.log] function to write logs to.
-class DevLogSink with LogSinkMixin {
+class DevLogSink extends LogSink {
   /// Creates a sink which writes its log-records with [developer.log].
   ///
   /// [filter] decides which log-records make it to [developer.log]. By
   /// default none are filtered out.
-  DevLogSink([this.filter = const LogFilter.none()]);
-
-  @override
-  final LogFilter filter;
+  DevLogSink([super.filter]);
 
   @override
   Future<void> write(LogRecord logRecord) {