+++ lib/src/sink/print_sink.dart
@@ -1,7 +1,6 @@
 import 'dart:async';
 import 'dart:convert';
 
-import 'package:fox_logging/src/filter/log_filter.dart';
 import 'package:fox_logging/src/formatter/log_record_formatter.dart';
 import 'package:fox_logging/src/sink/log_sink.dart';
 import 'package:logging/logging.dart';
@@ -9,7 +8,7 @@ import 'package:logging/logging.dart';
 /// A [LogSinkMixin] which uses the [print] function to write logs to.
 ///
 /// [formatter] is used to format [LogRecord] before printing it.
-class PrintSink with LogSinkMixin {
+class PrintSink extends LogSink {
   /// Creates a sink which writes its log-records with [print].
   ///
   /// [formatter] is used to format a [LogRecord] before printing it.
@@ -17,7 +16,7 @@ class PrintSink with LogSinkMixin {
   /// filtered out.
   PrintSink(
     this.formatter, [
-    this.filter = const LogFilter.none(),
+    super.filter,
   ]);
 
   /// Splits a formatted log-record into the lines which are printed
@@ -26,9 +25,6 @@ class PrintSink with LogSinkMixin {
 
   /// Used to format [LogRecord] before printing it.
   final LogRecordFormatter formatter;
-
-  @override
-  final LogFilter filter;
 
   @override
   Future<void> write(LogRecord logRecord) {